refactor: 리그 드롭다운 위치 수정
